+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `InputChips` and `SelectMulti` overflow when a fixed height is used
+- `InputChips` and `SelectMulti` long values breaking out of the input
 - `Select`/`SelectMulti` list closing when trying to scroll by dragging the scrollbar
 - `SelectMulti` list height now adjusts as needed when chips are added
 - `Select` getting text highlighted in a focus trap when not filterable